Fix deadlock if SERCOM RX gets ahead of USB
In this state (reachable by commenting out the uart_to_usb_consumer
release()), the SERCOM would receive a byte, but since the BBQueue was
full the ISR wouldn't read the byte out.  This lead to the RXC interrupt
continually firing, and since it's higher prio than USB the error never
cleared.